refactor-#29: move customer entity to entities folder
