+++ src/customer/customer.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { CommandBus, QueryBus } from '@nestjs/cqrs';
+
 import {
   CreateCustomerProfileCommand,
   RemoveCustomerProfileCommand,
@@ -7,7 +8,7 @@ import {
 } from './commands/impl';
 import { CreateCustomerProfileDto } from './dto/create-customer-profile.dto';
 import { GetCustomerProfileDto } from './dto/get-customer-profile.dto';
-import { CustomerEntity } from './repositories/customer.entity';
+import { CustomerEntity } from './entities';
 import { GetCustomerProfileQuery } from './queries/impl';
 import { RemoveCustomerProfileDto } from './dto/remove-customer-profile.dto';
 import { UpdateCustomerProfileDto } from './dto/update-customer-profile.dto';
