+++ src/customer/entities/customer.entity.ts
