+++ src/customer/repositories/customer.repository.ts
@@ -1,5 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { CustomerEntity } from './customer.entity';
+
+import { CustomerEntity } from '../entities';
 
 @EntityRepository(CustomerEntity)
 export class CustomerRepository extends Repository<CustomerEntity> {}
